Split files, fixed bug
