+++ Server/Helpers/IPlayerStateRequestsAdapter.cs
@@ -1,0 +1,9 @@
+﻿using Server.Models;
+
+namespace Server.Helpers
+{
+    interface IPlayerStateRequestsAdapter
+    {
+        GameState GetState();
+    }
+}
