+++ Server/Helpers/PlayerStateRequests.cs
@@ -27,24 +27,4 @@ namespace Server.Helpers
             return HttpRequests.GetRequest(_player.IpAddress + getEndpoint);
         }
     }
-
-    interface IPlayerStateRequestsAdapter
-    {
-        GameState GetState();
-    }
-
-    public class PlayerStateRequestsAdapter : IPlayerStateRequestsAdapter
-    {
-        PlayerStateRequests _playerStateRequests;
-
-        public PlayerStateRequestsAdapter(Player player)
-        {
-            _playerStateRequests = new PlayerStateRequests(player);
-        }
-
-        public GameState GetState()
-        {
-            return _playerStateRequests.GetState().Deserialize<GameState>(;
-        }
-    }
 }
